services/preprocessing-worker/requirements.txt
fastapi==0.104.1
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0